pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.191.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "openai-agents>=0.10.3",
    "python-dotenv>=1.2.2",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
]
